+++ packages/expo-cli/src/credentials/views/AndroidCredentials.ts
@@ -1,6 +1,5 @@
 import isEmpty from 'lodash/isEmpty';
 
-import CommandError from '../../CommandError';
 import log from '../../log';
 import prompt from '../../prompt';
 import { displayAndroidAppCredentials } from '../actions/list';